refactor(core): extract ContextAssembler as standalone testable struct (#2904)
Extract context assembly logic from Agent into a dedicated ContextAssembler
struct in crates/zeph-core/src/agent/context/assembler.rs.

ContextAssembler::gather() takes an explicit ContextAssemblyInput<'_> with
all borrowed fields and returns a PreparedContext — no access to Agent state.
All fetch_* helpers run concurrently via FuturesUnordered. Agent::prepare_context
becomes a ~30 LOC thin wrapper; apply_prepared_context handles injection,
sanitization, trimming, and session digest.

Closes #2904
