+++ crates/zeph-core/src/agent/context/mod.rs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
+pub(super) mod assembler;
 mod assembly;
 mod summarization;
